Relocated the Interfaces folder to be shared among platforms. Cleaned up accelerometer. Resolved warnings
