+++ Platform/STM/F4/CAN/bxcan_stmf4.hpp
@@ -21,7 +21,7 @@
 #include "can.h"
 
 // DFR Custom Dependencies
-#include "../../Interfaces/ican.hpp"
+#include "../../../Interfaces/ican.hpp"
 
 
 namespace platform {
