+++ Platform/STM/F4/GPIO/gpio_stmf4.hpp
@@ -16,7 +16,7 @@
 #include "gpio.h"
 
 // DFR Custom Dependencies
-#include "igpio.hpp"
+#include "../../../Interfaces/igpio.hpp"
 
 
 namespace platform {
